all methods inside an object with hooks
